Ambient mode: Adapt low-power-mode to allow for actual ambient mode. Refactor the low power mode to actually work in low power mode(deep sleep). There are some challenges when it comes to transitions (i.e. screen on -> screen lpm). There are multiple parts responsible for setting the next screen state.
Depending on what state the screen is the following may happen:
- Default, screen on (currently watchface is shown) -> screen lpm.
- In the `long timeout`(from asteroid-launcher) state, the screen may first go from screen on -> screen dim -> screen lpm
- When the screen is off and a notification arrives the screen isn't actually in the on state, so there is a change needed in tklock.c to fix that.

Another part that is very important is the ambient light sensor module.
There is also the filter-brightness-als module. This module is responsible for changing the range of the brightness depending on the currently set brightness and the ambient light sensor value.
So if the brightness is set to 0% the brightness profile use is 0, for 100% the profile 20 is used. These profiles are defined in `als-defaults.ini`.
For LPM/ambient mode a special profiles set is used (LPM).
In LPM mode we want to limit the maximum brightness.
So a default profile is used that limits the maximum brightness to 28%. If the als reports a lower value then a lower screen brightness is also possible.
An important side note is that, the als mode is enabled even if no such sensor exist. For those devices, the als value of 400 is used.
This value also signifies the maximum brightness in LPM mode.
Because the minimum brightness can vary between devices it is suggested to create alternative profiles for such devices.